Added feature histograms. Added PCA visualizations (see details).
PCA: Since the nan handling strategy is not finalized yet, only the clean (non-nan) part of the dataset, which is roughly 80 percent of the data, is used to find the principal components. An important note is that the percent null is similar for both failed and healthy drives, so the class ratio is somewhat maintained.
TSNE: sklearn tsne is very slow for the size of the dataset we have. However, some parallel implementations exist and dask can also be used as a parallel backend for sklearn. This also is very slow and will be looked into soon.