Added initial version of survival analysis ipynb
So far it contains a cox ph model trained on "time-collapsed" data (i.e., no time variation in covariates).
It also contains a setup for using coxph model with time varying covariates. The "covariate-update" interval is chosen to be 1 day to begin with because it is relatively simple to transform the input data to the required form. WARNING: model training not fully functioning because lifelines+dask do not appear to be directly compatible